requirements.txt
python-slugify==8.0.3
